% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{boxplot_DRG}
\alias{boxplot_DRG}
\title{Title: Create a Boxplot by DRG}
\usage{
boxplot_DRG(data, col_name)
}
\arguments{
\item{col_name}{A string.Takes in "covered_charges" for Average Covered Charges. Takes in "total_payments" for Average Total Payments. Takes in "medicare_payments" for Average Medicare Payments}
}
\value{
boxplot generated from ggplot2.
}
\description{
This function takes one string and returns a boxplot
}
\examples{
boxplot_DRG("covered_charges")  # returns boxplot by DRG
boxplot_DRG("total_payments")  # returns boxplot by DRG
}
