% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{DERG}
\format{
A data frame with 163065 rows and 12 columns.
}
\source{
source("~/Desktop/Data Science/lab/lab2-lab2group4/group4/R/data.R")
}
\usage{
data
}
\description{
DRG
}
\keyword{datasets}
